Moving CardInputHook code into CardInput. Deleting CardInputHook and CardInput > handlers
