+++ packages/lib/src/components/Card/components/CardInput/CardInput.tsx
@@ -1,330 +1,389 @@
-import { Component, h, createRef } from 'preact';
-import Address from '../../../internal/Address';
-import CardFields from './components/CardFields';
-import CardHolderName from './components/CardHolderName';
-import Installments from './components/Installments';
-import KCPAuthentication from './components/KCPAuthentication';
-import LoadingWrapper from '../../../internal/LoadingWrapper/LoadingWrapper';
-import StoredCardFields from './components/StoredCardFields';
-import SecuredFieldsProvider from '../../../../components/internal/SecuredFields/SecuredFieldsProvider';
-import StoreDetails from '../../../internal/StoreDetails';
+import { h } from 'preact';
+import { useState, useEffect, useRef, useMemo } from 'preact/hooks';
+import SecuredFieldsProvider, { SFPState } from '../../../internal/SecuredFields/SecuredFieldsProvider';
 import defaultProps from './defaultProps';
 import defaultStyles from './defaultStyles';
-import getImage from '../../../../utils/get-image';
-import handlers from './handlers';
-import processBinLookup from './processBinLookup';
 import styles from './CardInput.module.scss';
-import { CardInputProps, CardInputState } from './types';
 import './CardInput.scss';
-import { BinLookupResponse } from '../../types';
-import { CVC_POLICY_REQUIRED } from '../../../internal/SecuredFields/lib/configuration/constants';
-import { objectsDeepEqual } from '../../../internal/SecuredFields/lib/utilities/commonUtils';
+import LoadingWrapper from '../../../internal/LoadingWrapper';
+import StoredCardFields from './components/StoredCardFields';
+import Installments from './components/Installments';
+import CardFields from './components/CardFields';
+import KCPAuthentication from './components/KCPAuthentication';
 import SocialSecurityNumberBrazil from '../../../Boleto/components/SocialSecurityNumberBrazil/SocialSecurityNumberBrazil';
+import StoreDetails from '../../../internal/StoreDetails';
+import Address from '../../../internal/Address/Address';
+import getImage from '../../../../utils/get-image';
+import { CardInputProps, CardInputValidState, CardInputErrorState, CardInputDataState } from './types';
+import { CVC_POLICY_REQUIRED } from '../../../internal/SecuredFields/lib/configuration/constants';
+import { BinLookupResponse } from '../../types';
+import { cardInputFormatters, cardInputValidationRules, getRuleByNameAndMode } from './validate';
+import CIExtensions from './extensions';
+import { CbObjOnFocus } from '../../../internal/SecuredFields/lib/types';
+import CardHolderName from './components/CardHolderName';
+import useForm from '../../../../utils/useForm';
 
-class CardInput extends Component<CardInputProps, CardInputState> {
-    private readonly validateCardInput;
-    private readonly handleFocus;
-    private readonly handleAddress;
-    private readonly handleHolderName;
-    private readonly handleInstallments;
-    private readonly handleKCPAuthentication;
-    private readonly handleCPF;
-    private readonly handleSecuredFieldsChange;
-    private readonly handleOnStoreDetails;
-    private readonly handleAdditionalDataSelection;
-    private readonly processBinLookup: typeof processBinLookup;
+function CardInput(props: CardInputProps) {
+    const sfp = useRef(null);
+    const billingAddressRef = useRef(null);
 
-    public state;
-    public props;
-    private setFocusOn;
-    private updateStyles;
-    private handleUnsupportedCard;
-    private sfp = createRef();
-    private billingAddressRef = createRef();
-    private kcpAuthenticationRef = createRef();
+    if (process.env.NODE_ENV === 'development') this.sfp = sfp;
 
-    constructor(props) {
-        super(props);
+    /**
+     * STATE HOOKS
+     */
+    const [errors, setErrors] = useState<CardInputErrorState>({});
+    const [valid, setValid] = useState<CardInputValidState>({
+        ...(props.holderNameRequired && { holderName: false })
+    });
+    const [data, setData] = useState<CardInputDataState>({
+        ...(props.hasHolderName && { holderName: props.data.holderName ?? '' })
+    });
 
-        this.state = {
-            status: 'ready',
-            errors: {},
-            valid: {
-                ...(this.props.holderNameRequired && { holderName: false })
-            },
-            data: {
-                ...(this.props.hasHolderName && {
-                    holderName: this.props.holderName || this.props.data.holderName || ''
-                })
-            },
-            ...(this.props.billingAddressRequired && {
-                billingAddress: {
-                    ...this.props.data.billingAddress
-                }
-            }),
-            isValid: false,
-            cvcPolicy: CVC_POLICY_REQUIRED,
-            hideDateForBrand: false,
-            focusedElement: '',
-            additionalSelectElements: [],
-            additionalSelectValue: '',
-            issuingCountryCode: null
-        };
+    const [focusedElement, setFocusedElement] = useState('');
+    const [isSfpValid, setIsSfpValid] = useState(false);
+    const [hideDateForBrand, setHideDateForBrand] = useState(false);
+    const [cvcPolicy, setCvcPolicy] = useState(CVC_POLICY_REQUIRED);
+    const [issuingCountryCode, setIssuingCountryCode] = useState(null);
 
-        this.validateCardInput = handlers.validateCardInput.bind(this);
-        this.handleFocus = handlers.handleFocus.bind(this);
-        this.handleAddress = handlers.handleAddress.bind(this);
-        this.handleHolderName = handlers.handleHolderName.bind(this);
-        this.handleInstallments = handlers.handleInstallments.bind(this);
-        this.handleKCPAuthentication = handlers.handleKCPAuthentication.bind(this);
-        this.handleCPF = handlers.handleCPF.bind(this);
-        this.handleSecuredFieldsChange = handlers.handleSecuredFieldsChange.bind(this);
-        this.handleOnStoreDetails = handlers.handleOnStoreDetails.bind(this);
-        this.handleAdditionalDataSelection = handlers.handleAdditionalDataSelection.bind(this);
-        this.processBinLookup = processBinLookup.bind(this);
-    }
+    const [additionalSelectElements, setAdditionalSelectElements] = useState([]);
+    const [additionalSelectValue, setAdditionalSelectValue] = useState('');
 
-    public static defaultProps = defaultProps;
+    const [storePaymentMethod, setStorePaymentMethod] = useState(false);
+    const [billingAddress, setBillingAddress] = useState(props.billingAddressRequired ? props.data.billingAddress : null);
+    const [showSocialSecurityNumber, setShowSocialSecurityNumber] = useState(false);
+    const [socialSecurityNumber, setSocialSecurityNumber] = useState('');
+    const [installments, setInstallments] = useState({ value: null });
 
-    public componentDidMount() {
-        this.setFocusOn = this.sfp.current.setFocusOn;
-        this.updateStyles = this.sfp.current.updateStyles;
-        this.handleUnsupportedCard = this.sfp.current.handleUnsupportedCard;
-    }
+    /**
+     * LOCAL VARS
+     */
+    const {
+        handleChangeFor,
+        triggerValidation,
+        data: formData,
+        valid: formValid,
+        errors: formErrors,
+        setSchema,
+        setData: setFormData,
+        setValid: setFormValid,
+        setErrors: setFormErrors
+    } = useForm<CardInputDataState>({
+        schema: [],
+        defaultData: props.data,
+        formatters: cardInputFormatters,
+        rules: cardInputValidationRules
+    });
 
-    componentDidUpdate(prevProps, prevState) {
+    const hasInstallments = !!Object.keys(props.installmentOptions).length;
+    const showAmountsInInstallments = props.showInstallmentAmounts ?? true;
+
+    const cardCountryCode: string = issuingCountryCode ?? props.countryCode;
+    const isKorea = cardCountryCode === 'kr'; // If issuingCountryCode or the merchant defined countryCode is set to 'kr'
+    const showKCP = props.configuration.koreanAuthenticationRequired && isKorea;
+
+    const showBrazilianSSN: boolean =
+        (showSocialSecurityNumber && props.configuration.socialSecurityNumberMode === 'auto') ||
+        props.configuration.socialSecurityNumberMode === 'show';
+
+    /**
+     * HANDLERS
+     */
+    const handleFocus = (e: CbObjOnFocus) => {
+        setFocusedElement(e.currentFocusObject);
+        e.focus === true ? props.onFocus(e) : props.onBlur(e);
+    };
+
+    const handleOnStoreDetails = (storeDetails: boolean): void => {
+        setStorePaymentMethod(storeDetails);
+    };
+
+    const handleInstallments = (installments): void => {
+        setInstallments(installments);
+    };
+
+    const handleAddress = address => {
+        setFormData('billingAddress', address.data);
+        setFormValid('billingAddress', address.isValid);
+        setFormErrors('billingAddress', address.errors);
+    };
+
+    const handleSecuredFieldsChange = (sfState: SFPState): void => {
         /**
-         * Validating every time there's a change in state
+         * Handling auto complete value for holderName (but only if the component is using a holderName field)
          */
-        if (
-            !objectsDeepEqual(prevState.billingAddress, this.state.billingAddress) ||
-            prevState.storePaymentMethod !== this.state.storePaymentMethod ||
-            prevState.socialSecurityNumber !== this.state.socialSecurityNumber ||
-            !objectsDeepEqual(prevState.installments, this.state.installments) ||
-            prevState.isSfpValid !== this.state.isSfpValid ||
-            prevState.cvcPolicy !== this.state.cvcPolicy ||
-            prevState.hideDateForBrand !== this.state.hideDateForBrand ||
-            prevState.brand !== this.state.brand ||
-            prevState.additionalSelectValue !== this.state.additionalSelectValue ||
-            // Covers changes to:
-            // - encryptedField values
-            // - KCP authentication
-            // - holder name
-            !objectsDeepEqual(prevState.data, this.state.data)
-        ) {
-            this.validateCardInput();
+        if (sfState.autoCompleteName) {
+            if (!props.hasHolderName) return;
+            const holderNameValidationFn = getRuleByNameAndMode('holderName', 'blur');
+            const acHolderName = holderNameValidationFn(sfState.autoCompleteName) ? sfState.autoCompleteName : null;
+            if (acHolderName) {
+                setFormData('holderName', acHolderName);
+                setFormValid('holderName', true); // only if holderName is valid does this fny get called - so we know it's valid and w/o error
+                setFormErrors('holderName', null);
+            }
+            return;
         }
-    }
 
-    public componentWillUnmount() {
-        this.sfp.current.destroy();
-        this.sfp = null;
-    }
+        setData({ ...data, ...sfState.data });
+        setErrors({ ...errors, ...sfState.errors });
+        setValid({ ...valid, ...sfState.valid });
 
-    public getChildContext() {
-        return { i18n: this.props.i18n };
-    }
+        setIsSfpValid(sfState.isSfpValid);
 
-    public setStatus(status) {
-        this.setState({ status });
-    }
+        // Values relating to /binLookup response
+        setCvcPolicy(sfState.cvcPolicy);
+        setShowSocialSecurityNumber(sfState.showSocialSecurityNumber);
+        setHideDateForBrand(sfState.hideDateForBrand);
+    };
 
-    public resetAdditionalSelectState() {
-        this.setState({
-            additionalSelectElements: [],
-            additionalSelectValue: ''
-        });
-    }
+    // Farm the handlers for binLookup related functionality out to another 'extensions' file
+    const extensions = useMemo(
+        () =>
+            CIExtensions(
+                props,
+                { sfp },
+                { additionalSelectElements, setAdditionalSelectElements, setAdditionalSelectValue, issuingCountryCode, setIssuingCountryCode }
+            ),
+        [additionalSelectElements, issuingCountryCode]
+    );
 
-    public showValidation() {
+    /**
+     * EXPECTED METHODS ON CARD.THIS
+     */
+    this.showValidation = () => {
         // Validate SecuredFields
-        this.sfp.current.showValidation();
+        sfp.current.showValidation();
 
-        // Validate holderName
-        if (this.props.holderNameRequired && !this.state.valid.holderName) {
-            this.setState(prevState => ({
-                errors: { ...prevState.errors, holderName: true }
-            }));
-        }
-
-        if (
-            ((this.state.showSocialSecurityNumber && this.props.configuration.socialSecurityNumberMode === 'auto') ||
-                this.props.configuration.socialSecurityNumberMode === 'show') &&
-            !this.state.valid.socialSecurityNumber
-        ) {
-            this.setState(prevState => ({
-                errors: { ...prevState.errors, socialSecurityNumber: true }
-            }));
-        }
+        // Validates holderName & SSN & KCP (taxNumber)
+        triggerValidation();
 
         // Validate Address
-        if (this.billingAddressRef?.current) this.billingAddressRef.current.showValidation();
+        if (billingAddressRef?.current) billingAddressRef.current.showValidation();
+    };
 
-        // Validate KCP authentication
-        if (this.kcpAuthenticationRef?.current) this.kcpAuthenticationRef.current.showValidation();
-    }
+    this.processBinLookupResponse = (binLookupResponse: BinLookupResponse, isReset: boolean) => {
+        extensions.processBinLookup(binLookupResponse, isReset);
+    };
 
-    public processBinLookupResponse(data: BinLookupResponse, isReset: boolean) {
-        const issuingCountryCode = data?.issuingCountryCode ? data.issuingCountryCode.toLowerCase() : null;
+    /**
+     * EFFECT HOOKS
+     */
+    useEffect(() => {
+        // componentDidMount
+        this.setFocusOn = sfp.current.setFocusOn;
+        this.updateStyles = sfp.current.updateStyles;
+        this.handleUnsupportedCard = sfp.current.handleUnsupportedCard;
 
-        this.setState({ issuingCountryCode }, () => {
-            this.processBinLookup(data, isReset);
+        // componentWillUnmount
+        return () => {
+            sfp.current.destroy();
+        };
+    }, []);
+
+    /**
+     * Handle form schema updates
+     */
+    useEffect(() => {
+        const newSchema = [
+            ...(props.hasHolderName ? ['holderName'] : []),
+            ...(showBrazilianSSN ? ['socialSecurityNumber'] : []),
+            ...(showKCP ? ['taxNumber'] : []),
+            ...(props.billingAddressRequired ? ['billingAddress'] : [])
+        ];
+        setSchema(newSchema);
+    }, [props.hasHolderName, showBrazilianSSN, showKCP]);
+
+    /**
+     * Handle updates from useForm
+     */
+    useEffect(() => {
+        setData({ ...data, holderName: formData.holderName ?? '', taxNumber: formData.taxNumber });
+
+        setSocialSecurityNumber(formData.socialSecurityNumber);
+
+        if (props.billingAddressRequired) setBillingAddress({ ...formData.billingAddress });
+
+        setValid({
+            ...valid,
+            holderName: props.holderNameRequired ? formValid.holderName : true,
+            // Setting value to false if it's falsy keeps in line with existing, expected behaviour
+            // - but there is an argument to allow 'undefined' as a value to indicate the non-presence of the field
+            socialSecurityNumber: formValid.socialSecurityNumber ? formValid.socialSecurityNumber : false,
+            taxNumber: formValid.taxNumber ? formValid.taxNumber : false,
+            billingAddress: formValid.billingAddress ? formValid.billingAddress : false
         });
-    }
 
-    render(
-        { countryCode, loadingContext, hasHolderName, hasCVC, installmentOptions, positionHolderNameOnTop, showInstallmentAmounts, configuration },
-        { status, cvcPolicy, hideDateForBrand, focusedElement, issuingCountryCode }
-    ) {
-        const hasInstallments = !!Object.keys(installmentOptions).length;
+        // Errors
+        setErrors({
+            ...errors,
+            holderName: props.holderNameRequired && !!formErrors.holderName ? formErrors.holderName : null,
+            socialSecurityNumber: showBrazilianSSN && !!formErrors.socialSecurityNumber ? formErrors.socialSecurityNumber : null,
+            taxNumber: showKCP && !!formErrors.taxNumber ? formErrors.taxNumber : null,
+            billingAddress: props.billingAddressRequired && !!formErrors.billingAddress ? formErrors.billingAddress : null
+        });
+    }, [formData, formValid, formErrors]);
 
-        // In the Drop-in the oneClick status may already have been decided, so give that priority
-        const isOneClick = this.props.oneClick || !!this.props.storedPaymentMethodId;
+    /**
+     * Main 'componentDidUpdate' handler
+     */
+    useEffect(() => {
+        const holderNameValid: boolean = valid.holderName;
 
-        const cardCountryCode: string = issuingCountryCode ?? countryCode;
+        const sfpValid: boolean = isSfpValid;
+        const addressValid: boolean = props.billingAddressRequired ? valid.billingAddress : true;
 
-        // If issuingCountryCode is set or the merchant defined countryCode is 'KR'
-        const isKorea = cardCountryCode === 'kr';
+        const koreanAuthentication: boolean = showKCP ? !!valid.taxNumber && !!valid.encryptedPassword : true;
 
-        const showAmountsInInstallments = showInstallmentAmounts ?? true;
+        const socialSecurityNumberValid: boolean = showBrazilianSSN ? !!valid.socialSecurityNumber : true;
 
-        // Brazilian socialSecurityNumber
-        const showBrazilianSSN: boolean =
-            (this.state.showSocialSecurityNumber && configuration.socialSecurityNumberMode === 'auto') ||
-            configuration.socialSecurityNumberMode === 'show';
+        const isValid: boolean = sfpValid && holderNameValid && addressValid && koreanAuthentication && socialSecurityNumberValid;
 
-        const cardHolderField = (
-            <CardHolderName
-                required={this.props.holderNameRequired}
-                placeholder={this.props.placeholders.holderName}
-                value={this.state.data.holderName}
-                error={!!this.state.errors.holderName}
-                isValid={!!this.state.valid.holderName}
-                onChange={this.handleHolderName}
-            />
-        );
+        props.onChange({
+            data,
+            valid,
+            errors,
+            isValid,
+            billingAddress,
+            additionalSelectValue,
+            storePaymentMethod,
+            socialSecurityNumber,
+            installments
+        });
+    }, [data, valid, errors, additionalSelectValue, storePaymentMethod, installments]);
 
-        return (
-            <SecuredFieldsProvider
-                ref={this.sfp}
-                {...this.props}
-                styles={{ ...defaultStyles, ...this.props.styles }}
-                koreanAuthenticationRequired={configuration.koreanAuthenticationRequired}
-                hasKoreanFields={!!(configuration.koreanAuthenticationRequired && countryCode === 'kr')}
-                onChange={this.handleSecuredFieldsChange}
-                onBrand={this.props.onBrand}
-                onFocus={this.handleFocus}
-                type={this.props.brand}
-                oneClick={isOneClick}
-                render={({ setRootNode, setFocusOn }, sfpState) => (
-                    <div
-                        ref={setRootNode}
-                        className={`adyen-checkout__card-input ${styles['card-input__wrapper']} adyen-checkout__card-input--${this.props
-                            .fundingSource ?? 'credit'}`}
-                    >
-                        {this.props.storedPaymentMethodId ? (
-                            <LoadingWrapper status={sfpState.status}>
-                                <StoredCardFields
-                                    {...this.props}
-                                    errors={sfpState.errors}
-                                    brand={sfpState.brand}
-                                    hasCVC={hasCVC}
-                                    cvcPolicy={cvcPolicy}
+    /**
+     * RENDER
+     */
+    const cardHolderField = (
+        <CardHolderName
+            required={props.holderNameRequired}
+            placeholder={props.placeholders.holderName}
+            value={data.holderName}
+            error={!!errors.holderName}
+            isValid={!!valid.holderName}
+            onChange={handleChangeFor('holderName', 'blur')}
+            onInput={handleChangeFor('holderName', 'input')}
+        />
+    );
+
+    const getInstallmentsComp = brand => (
+        <Installments
+            amount={props.amount}
+            brand={brand}
+            installmentOptions={props.installmentOptions}
+            onChange={handleInstallments}
+            type={showAmountsInInstallments ? 'amount' : 'months'}
+        />
+    );
+
+    return (
+        <SecuredFieldsProvider
+            ref={sfp}
+            {...props}
+            styles={{ ...defaultStyles, ...props.styles }}
+            koreanAuthenticationRequired={props.configuration.koreanAuthenticationRequired}
+            hasKoreanFields={!!(props.configuration.koreanAuthenticationRequired && props.countryCode === 'kr')}
+            onChange={handleSecuredFieldsChange}
+            onBrand={props.onBrand}
+            onFocus={handleFocus}
+            type={props.brand}
+            render={({ setRootNode, setFocusOn }, sfpState) => (
+                <div
+                    ref={setRootNode}
+                    className={`adyen-checkout__card-input ${styles['card-input__wrapper']} adyen-checkout__card-input--${props.fundingSource ??
+                        'credit'}`}
+                >
+                    {props.storedPaymentMethodId ? (
+                        <LoadingWrapper status={sfpState.status}>
+                            <StoredCardFields
+                                {...props}
+                                errors={sfpState.errors}
+                                brand={sfpState.brand}
+                                hasCVC={props.hasCVC}
+                                cvcPolicy={cvcPolicy}
+                                onFocusField={setFocusOn}
+                                focusedElement={focusedElement}
+                                status={sfpState.status}
+                                valid={sfpState.valid}
+                            />
+
+                            {hasInstallments && getInstallmentsComp(sfpState.brand)}
+                        </LoadingWrapper>
+                    ) : (
+                        <LoadingWrapper status={sfpState.status}>
+                            {props.hasHolderName && props.positionHolderNameOnTop && cardHolderField}
+
+                            <CardFields
+                                {...props}
+                                brand={sfpState.brand}
+                                focusedElement={focusedElement}
+                                onFocusField={setFocusOn}
+                                hasCVC={props.hasCVC}
+                                cvcPolicy={cvcPolicy}
+                                hideDateForBrand={hideDateForBrand}
+                                errors={sfpState.errors}
+                                valid={sfpState.valid}
+                                dualBrandingElements={additionalSelectElements.length > 0 && additionalSelectElements}
+                                dualBrandingChangeHandler={extensions.handleDualBrandSelection}
+                                dualBrandingSelected={additionalSelectValue}
+                            />
+
+                            {props.hasHolderName && !props.positionHolderNameOnTop && cardHolderField}
+
+                            {showKCP && (
+                                <KCPAuthentication
                                     onFocusField={setFocusOn}
                                     focusedElement={focusedElement}
-                                    status={sfpState.status}
-                                    valid={sfpState.valid}
+                                    encryptedPasswordState={{
+                                        data: sfpState.encryptedPassword,
+                                        valid: sfpState.valid ? sfpState.valid.encryptedPassword : false,
+                                        errors: sfpState.errors ? sfpState.errors.encryptedPassword : false
+                                    }}
+                                    value={data.taxNumber}
+                                    error={!!errors.taxNumber}
+                                    isValid={!!valid.taxNumber}
+                                    onChange={handleChangeFor('taxNumber', 'blur')}
+                                    onInput={handleChangeFor('taxNumber', 'input')}
                                 />
+                            )}
 
-                                {hasInstallments && (
-                                    <Installments
-                                        amount={this.props.amount}
-                                        brand={sfpState.brand}
-                                        installmentOptions={installmentOptions}
-                                        onChange={this.handleInstallments}
-                                        type={showAmountsInInstallments ? 'amount' : 'months'}
+                            {showBrazilianSSN && (
+                                <div className="adyen-checkout__card__socialSecurityNumber">
+                                    <SocialSecurityNumberBrazil
+                                        onChange={handleChangeFor('socialSecurityNumber', 'blur')}
+                                        onInput={handleChangeFor('socialSecurityNumber', 'input')}
+                                        error={errors?.socialSecurityNumber}
+                                        valid={valid?.socialSecurityNumber}
+                                        data={socialSecurityNumber}
                                     />
-                                )}
-                            </LoadingWrapper>
-                        ) : (
-                            <LoadingWrapper status={sfpState.status}>
-                                {hasHolderName && positionHolderNameOnTop && cardHolderField}
+                                </div>
+                            )}
 
-                                <CardFields
-                                    {...this.props}
-                                    brand={sfpState.brand}
-                                    focusedElement={focusedElement}
-                                    onFocusField={setFocusOn}
-                                    hasCVC={hasCVC}
-                                    cvcPolicy={cvcPolicy}
-                                    hideDateForBrand={hideDateForBrand}
-                                    errors={sfpState.errors}
-                                    valid={sfpState.valid}
-                                    dualBrandingElements={this.state.additionalSelectElements.length > 0 && this.state.additionalSelectElements}
-                                    dualBrandingChangeHandler={this.handleAdditionalDataSelection}
-                                    dualBrandingSelected={this.state.additionalSelectValue}
+                            {props.enableStoreDetails && <StoreDetails onChange={handleOnStoreDetails} />}
+
+                            {hasInstallments && getInstallmentsComp(sfpState.brand)}
+
+                            {props.billingAddressRequired && (
+                                <Address
+                                    label="billingAddress"
+                                    data={billingAddress}
+                                    onChange={handleAddress}
+                                    allowedCountries={props.billingAddressAllowedCountries}
+                                    requiredFields={props.billingAddressRequiredFields}
+                                    ref={billingAddressRef}
                                 />
+                            )}
+                        </LoadingWrapper>
+                    )}
 
-                                {hasHolderName && !positionHolderNameOnTop && cardHolderField}
-
-                                {configuration.koreanAuthenticationRequired && isKorea && (
-                                    <KCPAuthentication
-                                        onFocusField={setFocusOn}
-                                        focusedElement={focusedElement}
-                                        encryptedPasswordState={{
-                                            data: sfpState.encryptedPassword,
-                                            valid: sfpState.valid ? sfpState.valid.encryptedPassword : false,
-                                            errors: sfpState.errors ? sfpState.errors.encryptedPassword : false
-                                        }}
-                                        ref={this.kcpAuthenticationRef}
-                                        onChange={this.handleKCPAuthentication}
-                                    />
-                                )}
-
-                                {showBrazilianSSN && (
-                                    <div className="adyen-checkout__card__socialSecurityNumber">
-                                        <SocialSecurityNumberBrazil
-                                            onChange={e => this.handleCPF(e, true)}
-                                            onInput={e => this.handleCPF(e)}
-                                            error={this.state.errors?.socialSecurityNumber}
-                                            valid={this.state.valid?.socialSecurityNumber}
-                                            data={this.state.socialSecurityNumber}
-                                        />
-                                    </div>
-                                )}
-
-                                {this.props.enableStoreDetails && <StoreDetails onChange={this.handleOnStoreDetails} />}
-
-                                {hasInstallments && (
-                                    <Installments
-                                        amount={this.props.amount}
-                                        brand={sfpState.brand}
-                                        installmentOptions={installmentOptions}
-                                        onChange={this.handleInstallments}
-                                        type={showAmountsInInstallments ? 'amount' : 'months'}
-                                    />
-                                )}
-
-                                {this.props.billingAddressRequired && (
-                                    <Address
-                                        label="billingAddress"
-                                        data={this.state.billingAddress}
-                                        onChange={this.handleAddress}
-                                        allowedCountries={this.props.billingAddressAllowedCountries}
-                                        requiredFields={this.props.billingAddressRequiredFields}
-                                        ref={this.billingAddressRef}
-                                    />
-                                )}
-                            </LoadingWrapper>
-                        )}
-
-                        {this.props.showPayButton &&
-                            this.props.payButton({ status, icon: getImage({ loadingContext, imageFolder: 'components/' })('lock') })}
-                    </div>
-                )}
-            />
-        );
-    }
+                    {props.showPayButton &&
+                        props.payButton({ status, icon: getImage({ loadingContext: props.loadingContext, imageFolder: 'components/' })('lock') })}
+                </div>
+            )}
+        />
+    );
 }
+
+CardInput.defaultProps = defaultProps;
 
 export default CardInput;
